+++ pocketutils/tools/all_tools.py
@@ -1,3 +1,4 @@
+from pocketutils.tools.loop_tools import LoopTools
 from pocketutils.tools.call_tools import CallTools
 from pocketutils.tools.common_tools import CommonTools
 from pocketutils.tools.console_tools import ConsoleTools
@@ -5,6 +6,7 @@ from pocketutils.tools.filesys_tools import FilesysTools
 from pocketutils.tools.numeric_tools import NumericTools
 from pocketutils.tools.path_tools import PathTools
 from pocketutils.tools.program_tools import ProgramTools
+from pocketutils.tools.reflection_tools import ReflectionTools
 from pocketutils.tools.string_tools import StringTools
 from pocketutils.tools.unit_tools import UnitTools
 
@@ -19,6 +21,8 @@ class Tools(
     StringTools,
     FilesysTools,
     UnitTools,
+    LoopTools,
+    ReflectionTools,
 ):
     """
     A collection of utility static functions.